% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygentags.R
\name{dummy}
\alias{dummy}
\title{Roxygen commands}
\usage{
dummy()
}
\description{
This is a dummy function who's purpose is to hold the useDynLib roxygen tag.
This tag will populate the namespace with compiled c++ functions upon package install.
}
