% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.genDA.R
\name{print.multiDA}
\alias{print.multiDA}
\title{Print Method for genDA}
\usage{
\method{print}{multiDA}(object, ...)
}
\arguments{
\item{object}{an object of class 'genDA' to print}

\item{...}{Any other variables which will be ignored.}
}
\description{
Prints summarised information from the genDA function.
}
