% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.genDA.R
\name{predict.genDA}
\alias{predict.genDA}
\alias{predict}
\title{Predict Method for genDA}
\usage{
\method{predict}{genDA}(object, new.y, newX = NULL, prior_beta = c(1,
  1), ...)
}
\arguments{
\item{object}{an object of class 'genDA'.}

\item{new.y}{A new data frame or matrix of response data. Must be numeric. Column response (family) types will be assumed to match that of the trained genDA object.}

\item{newX}{A new data frame or matrix of known covariate data.}

\item{prior_beta}{A vector of parameters for the beta prior for class prediction. Defaults to c(1,1) (Uninformative prior)}

\item{...}{not used.}
}
\value{
An object including the following components:

 \item{class }{a vector of predicted class labels}
 \item{prob_class }{a matrix of class membership probabilities}
}
\description{
Obtains new class predictions from a fitted genDA object.
}
\author{
Sarah Romanes <sarah.romanes@sydney.edu.au>
}
